Add parent classes into query requests & responses, folderise #79
