+++ src/Message/Query/QueryBatchResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Omnipay\AuthorizeNet\Message;
+namespace Omnipay\AuthorizeNet\Message\Query;
 
 use Omnipay\AuthorizeNet\Model\CardReference;
 use Omnipay\AuthorizeNet\Model\TransactionReference;
@@ -11,7 +11,7 @@ use Omnipay\Common\Message\AbstractResponse;
 /**
  * Authorize.Net AIM Response
  */
-class QueryBatchResponse extends AbstractResponse
+class QueryBatchResponse extends AbstractQueryResponse
 {
     /**
      * For Error codes: @see https://developer.authorize.net/api/reference/responseCodes.html
@@ -51,32 +51,5 @@ class QueryBatchResponse extends AbstractResponse
     {
         $result = $this->xml2array($this->data->batchList, true);
         return $result['batchList'][0]['batch'];
-    }
-
-    /**
-     * http://bookofzeus.com/articles/convert-simplexml-object-into-php-array/
-     *
-     * Convert a simpleXMLElement in to an array
-     *
-     * @todo this is duplicated from CIMAbstractResponse. Put somewhere shared.
-     *
-     * @param \SimpleXMLElement $xml
-     *
-     * @return array
-     */
-    public function xml2array(\SimpleXMLElement $xml)
-    {
-        $arr = array();
-        foreach ($xml as $element) {
-            $tag = $element->getName();
-            $e = get_object_vars($element);
-            if (!empty($e)) {
-                $arr[$tag][] = $element instanceof \SimpleXMLElement ? $this->xml2array($element) : $e;
-            } else {
-                $arr[$tag] = trim($element);
-            }
-        }
-
-        return $arr;
     }
 }
