+++ src/Message/Query/QueryDetailResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Omnipay\AuthorizeNet\Message;
+namespace Omnipay\AuthorizeNet\Message\Query;
 
 use Omnipay\AuthorizeNet\Model\CardReference;
 use Omnipay\AuthorizeNet\Model\TransactionReference;
@@ -12,7 +12,7 @@ use Omnipay\Omnipay;
 /**
  * Authorize.Net AIM Response
  */
-class QueryDetailResponse extends AbstractResponse
+class QueryDetailResponse extends AbstractQueryResponse
 {
     protected $planResponse;
 
@@ -240,32 +240,5 @@ class QueryDetailResponse extends AbstractResponse
     public function getTransactionDate()
     {
         return $this->transaction['submitTimeUTC'];
-    }
-
-    /**
-     * http://bookofzeus.com/articles/convert-simplexml-object-into-php-array/
-     *
-     * Convert a simpleXMLElement in to an array
-     *
-     * @todo this is duplicated from CIMAbstractResponse. Put somewhere shared.
-     *
-     * @param \SimpleXMLElement $xml
-     *
-     * @return array
-     */
-    public function xml2array(\SimpleXMLElement $xml)
-    {
-        $arr = array();
-        foreach ($xml as $element) {
-            $tag = $element->getName();
-            $e = get_object_vars($element);
-            if (!empty($e)) {
-                $arr[$tag][] = $element instanceof \SimpleXMLElement ? $this->xml2array($element) : $e;
-            } else {
-                $arr[$tag] = trim($element);
-            }
-        }
-
-        return $arr;
     }
 }
