+++ src/Message/Query/QueryResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Omnipay\AuthorizeNet\Message;
+namespace Omnipay\AuthorizeNet\Message\Query;
 
 use Omnipay\AuthorizeNet\Model\CardReference;
 use Omnipay\AuthorizeNet\Model\TransactionReference;
@@ -38,32 +38,5 @@ class QueryResponse extends QueryBatchResponse
             }
         }
         return $result;
-    }
-
-    /**
-     * http://bookofzeus.com/articles/convert-simplexml-object-into-php-array/
-     *
-     * Convert a simpleXMLElement in to an array
-     *
-     * @todo this is duplicated from CIMAbstractResponse. Put somewhere shared.
-     *
-     * @param \SimpleXMLElement $xml
-     *
-     * @return array
-     */
-    public function xml2array(\SimpleXMLElement $xml)
-    {
-        $arr = array();
-        foreach ($xml as $element) {
-            $tag = $element->getName();
-            $e = get_object_vars($element);
-            if (!empty($e)) {
-                $arr[$tag][] = $element instanceof \SimpleXMLElement ? $this->xml2array($element) : $e;
-            } else {
-                $arr[$tag] = trim($element);
-            }
-        }
-
-        return $arr;
     }
 }
